:fire: :sparkles:
Removed unnecessary files for the computation of the Hermite polynomial kernel matrix, work is needed on the Gegenbauer one
