+++ shermite.py
@@ -1,14 +1,20 @@
 import numpy as np
-from numba import njit
-from sklearn.svm import SVC
-from sklearn.preprocessing import MinMaxScaler
-from padierna_modules.plots import plot_svc_decision_function
+import pandas as pd
 import matplotlib.pyplot as plt
-from multiprocessing import Pool
-from itertools import repeat
-from sklearn.model_selection import StratifiedKFold
+from sklearn.svm import SVC
+from sklearn.datasets.samples_generator import make_circles
+from padierna_modules.plots import plot_svc_decision_function
+from numba import njit
 
 
+# SELECCIÓN DE PROCESAMIENTO
+# *********************************************
+analizarFourclass = True
+graficarPolinomios = False
+calcularGram = True
+
+# POLINOMIO HERMITE
+# *******************************************
 @njit
 def hermite(x, n):
 
@@ -30,106 +36,119 @@ def hermite(x, n):
         return resultado
 
 
-# @njit
-# def special_hermite(x, z, n):
+# MOSTRANDO POLINOMIOS DE S-HERMITE PARA VALIDAR LA FUNCIÓN H(x_i,n) ESCALADA
+# *********************************************
+if graficarPolinomios:
+    plt.figure()
+    t = np.arange(-1, 1.1, 0.1)  # Rango de prueba (test)
+    for i in range(1, 6):
+        plt.plot(t, hermite(t, i) * 2 ** (-i), label="Grado " + str(i))
+        plt.legend()
+        plt.title("Polinomios Ortogonales de s-Hermite")
 
-#     prod_1 = hermite(x, n) * hermite(z, n)
-#     # prod_2 = np.exp(-0.5 * (x ** 2 + z ** 2))
-#     # result = prod_1 * prod_2 / 2.0 ** (2.0 * n)
-#     # result = prod_1
-
-#     return prod_1
-
-
-# @memory.cache
-def kernel_special_hermite(X, degree=2):
+# KERNEL S-HERMITE:
+# ********************************************
+def sHerm_kernel(X, degree=2):
 
     X_gram = np.zeros((X.shape[0], X.shape[0]))
 
     for l, x in enumerate(X):
         for m, z in enumerate(X):
-            # Aprovechar la simetría de la matriz
+            summ, mult, i, j = 1.0, 1.0, 0, 0
+            # Saltarse la triangular inferior
             if l > m:
                 continue
-            mult = 1.0
-            sum_res = 1.0
-            for j, u in zip(x, z):
-                sum_res += sum(
-                    hermite(j, n) * hermite(u, n) / (2.0 ** (2.0 * degree))
-                    for n in range(1, degree + 1)
-                )
-            mult *= sum_res
+            # Computar la matriz con kernel Hermite
+            while i < x.size and j < z.size:
+                if i == j:
+                    summ = 1.0
+                    for k in range(1, degree + 1):
+                        summ += (
+                            hermite(x[i], k) * hermite(z[j], k) / (2.0 ** (2 * degree))
+                        )
+                    mult *= summ
+                    i += 1
+                    j += 1
+                else:
+                    if i > j:
+                        j += 1
+                    else:
+                        i += 1
             X_gram[l, m] = mult
-    # Completar la matriz simétrica con la parte superior de la matriz triangular
-    X_gram = np.triu(X_gram) + np.triu(X_gram, 1).T
+        # Completar la matriz simétrica con la parte superior de la matriz triangular
+        X_gram = np.triu(X_gram) + np.triu(X_gram, 1).T
 
-    return X_gram
+    return np.array(X_gram)
 
 
-fourclass = np.genfromtxt("fourclass1.csv", delimiter=",", skip_header=1)
-X, y = fourclass[:, 0:2], fourclass[:, 2]
-# Escalar a -1 y 1
-# X = MinMaxScaler(feature_range=(-1.0, 1.0)).fit_transform(X)
-# PARÁMETROS ÓPTIMOS (Del artículo sobre Gegenbauer)
+# SELECCIÓN DEL DATASET
 # *********************************************
-C_sH, degree_sH = 25.20, 6  # s-Herm - Fourclass
+if analizarFourclass:
+    fourclass = pd.read_csv("fourclass1.csv")
+    X, Y = (fourclass.iloc[:, 0:2]), fourclass["Y"]
+    # PARÁMETROS ÓPTIMOS (Del artículo sobre Gegenbauer)
+    # *********************************************
+    C_rbf, gamma_rbf = 30.42, 3.82  # RBF    - Fourclass
+    C_sH, degree_sH = 25.20, 6  # s-Herm - Fourclass
+else:
+    X, Y = make_circles(100, factor=0.3, noise=0.1)
+    # PARÁMETROS DE PRUEBA
+    # *********************************************
+    C_rbf, gamma_rbf = 100, 0.1  # RBF     - Dataset X
+    C_sH, degree_sH = 100, 2  # s-Herm - Dataset X
+
 # VERIFICANDO MATRIZ GRAMIANA DE S-HERM
 # *********************************************
-X_gram = kernel_special_hermite(X, degree=degree_sH)
-print(X_gram)
-print("\n****VERIFICANDO MATRIZ GRAM*****")
-print(type(X_gram))
-print("Gram Max =", X_gram.max(), "Gram min =", X_gram.min())
-NANs = np.argwhere(np.isnan(X_gram))
-print("Valores tipo NAN: ", NANs)
-# # ENTRENANDO MSV CON RBF y S-HERM.
-# dict_hermite = {"C": C_sH, "kernel": "precomputed"}
+X = np.array(X)
+if calcularGram:
+    X_gram = sHerm_kernel(X, degree=degree_sH)
+    print(X_gram)
+    print("\n****VERIFICANDO MATRIZ GRAM*****")
+    print(type(X_gram))
+    print("Gram Max =", X_gram.max(), "Gram min =", X_gram.min())
+    NANs = np.argwhere(np.isnan(X_gram))
+    print("Valores tipo NAN: ", NANs)
 
-# svc_hermite = SVC(**dict_hermite)
-
-# rscv = StratifiedKFold(n_splits=10)
-
-# # def train_model(params):
-# def train_model(model, x, y):
-
-#     x = np.array(x)
-#     y = np.array(y)
-#     result = []
-#     for i, j in zip(x, y):
-#         gram = kernel_special_hermite(i, degree=degree_sH)
-#         vectors = model.fit(gram, j).support_
-#         result.append(len(vectors))
-
-#     return result
-
-
-# svc_1 = []
-# x_training = 0
-# y_training = []
-
-# for train_idx, _ in rscv.split(X, y):
-
-#     x_training = len(X[train_idx])
-#     X_gram = kernel_special_hermite(X[train_idx], degree=degree_sH)
-#     svc_1.append(len(svc_hermite.fit(X_gram, y[train_idx]).support_))
-
-# # x_split = np.split(np.array(x_training), 5)
-# # y_split = np.split(np.array(y_training), 5)
-
-# # with Pool(4) as pool:
-
-# #     svc_1.append(
-# #         pool.map_async(
-# #             train_model,
-# #             zip(
-# #                 repeat(svc_hermite),
-# #                 [i for i in x_split[:4]],
-# #                 [i for i in y_split[:4]],
-# #             ),
-# #         ).get()
-# #     )
-
-# # svc_1 = np.array(svc_1).ravel()
-# # svc_1 = np.append(svc_1, train_model((svc_hermite, x_split[-1], y_split[-1])))
-# psv = np.array(svc_1) * 100.0 / x_training
-# print(psv.mean(), psv.std())
+# # ENTRENANDO Y GRAFICANDO MSV CON RBF y S-HERM.
+# plt.figure()
+# clf = SVC(kernel="rbf", C=C_rbf, gamma=gamma_rbf).fit(X, Y)
+# clf2 = SVC(kernel=sHerm_kernel, C=C_sH, degree=degree_sH).fit(X, Y)
+# plt.subplot(1, 2, 1)
+# plt.title(
+#     "MSV-RBF: C="
+#     + str(C_rbf)
+#     + ",gamma="
+#     + str(gamma_rbf)
+#     + ",SVs="
+#     + str(len(clf.support_))
+# )
+# plt.scatter(X[:, 0], X[:, 1], c=Y, s=50, cmap="cool")
+# plot_svc_decision_function(clf, X, plot_support=True)
+# plt.subplot(1, 2, 2)
+# plt.title(
+#     "MSV-sHerm: C="
+#     + str(C_sH)
+#     + " n="
+#     + str(degree_sH)
+#     + ",SVs="
+#     + str(len(clf2.support_))
+# )
+# plt.scatter(X[:, 0], X[:, 1], c=Y, s=50, cmap="cool")
+# clf2.support_vectors_ = X[np.array(clf2.support_), :]
+# plot_svc_decision_function(clf2, X, plot_support=True, customKernel=True)
+# print("\n*************************************************************")
+# print("RESULTADOS DE MODELOS RBF Y S-HERM")
+# print("***************************************************************")
+# print(
+#     "Vectores Soporte (VS) RBF:\t" + str(len(clf.support_)),
+#     "\ts-Herm: ",
+#     str(len(clf2.support_)),
+# )
+# print(
+#     "PSV: {0} {1}".format(
+#         len(clf.support_) * 100.0 / X.shape[0], len(clf2.support_) * 100.0 / X.shape[0]
+#     )
+# )
+# print("VS por Clase RBF:\t\t" + str(clf.n_support_), "\ts-Herm: ", str(clf2.n_support_))
+# print("Indices VS RBF:\t\t\t" + str(clf.support_), "\ts-Herm: ", str(clf2.support_))
+# plt.show()
